+++ src/test/java/com/wanted/preonboarding/security/annotation/CustomWebMvcTest.java
@@ -1,4 +1,4 @@
-package com.wanted.preonboarding.security;
+package com.wanted.preonboarding.security.annotation;
 
 import org.springframework.boot.autoconfigure.security.oauth2.client.servlet.OAuth2ClientAutoConfiguration;
 import org.springframework.boot.autoconfigure.security.oauth2.resource.servlet.OAuth2ResourceServerAutoConfiguration;
